+++ owlbot.py
@@ -1,6 +1,4 @@
-#!/bin/bash
-
-# Copyright 2021 Google LLC
+# Copyright 2018 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,16 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-set -eo pipefail
+import synthtool.languages.node as node
 
-if [[ -z "$CREDENTIALS" ]]; then
-  # if CREDENTIALS are explicitly set, assume we're testing locally
-  # and don't set NPM_CONFIG_PREFIX.
-  export NPM_CONFIG_PREFIX=${HOME}/.npm-global
-  export PATH="$PATH:${NPM_CONFIG_PREFIX}/bin"
-  cd $(dirname $0)/../..
-fi
-
-npm install
-npm install --no-save @google-cloud/cloud-rad@^0.2.5
-npx @google-cloud/cloud-rad
+node.owlbot_main(templates_excludes=["LICENSE", "README.md"])