+++ src/warnings.ts
@@ -18,7 +18,8 @@ import {isBrowser} from './isbrowser';
 
 const emittedWarnings = new Set<string>();
 
-export function warn(code: string, message: string) {
+// warnType is the type of warning (e.g. 'DeprecationWarning', 'ExperimentalWarning', etc.)
+export function warn(code: string, message: string, warnType?: string) {
   // Only show a given warning once
   if (emittedWarnings.has(code)) {
     return;
@@ -27,6 +28,10 @@ export function warn(code: string, message: string) {
 
   if (isBrowser()) {
     console.warn(message);
+  } else if (typeof warnType !== 'undefined') {
+    process.emitWarning(message, {
+      type: warnType,
+    });
   } else {
     process.emitWarning(message);
   }